configs: Update scripts and spec file with layout changes
With the configs in the new place, update the scripts to find
them and utilize the base-generic and generic heirarchy to apply
configs and overrides.

Implement new process_configs.sh script that post-process the
config changes in a simple script and remove those commands
from the spec file.  Add option flags to preserve functionality.

config_generation is a simple rename of baseconfig -> generic and
debugconfig -> debug and arm64 -> aarach64.

build_configs.sh is modified to find configs in generic and base-generic and
then apply base-generic first and if any generic files, apply those next.  The
generic directory is used as an overrides and is expected to be empty for
Fedora initially.

kernel.spec is modified to use process_configs.sh instead of all the
commands in the spec file.  Enabled spec options are translated to
script options.  The config manipulation is moved to be grouped
with all config manipulation commands.  This makes 'cd configs/' simpler.
Now all config scripts and executiion are done in configs/ directory.

v1 -> v2:
* the scripts were not working with SUBARCH correctly
* checkoptions was using wrong comparison file
* passing wrong kernel version to process_configs in spec file

v2 -> v3:
(incorporate Laura A's feedback)
* update README.txt
* fix build_configs.sh warnings
* Output info message on listnewconfig failure

+++ configs/build_configs.sh
@@ -4,12 +4,32 @@
 # and debug to form the necessary $PACKAGE_NAME<version>-<arch>-<variant>.config
 # files for building RHEL kernels, based on the contents of a control file
 
-PACKAGE_NAME=kernel # defines the package name used
+PACKAGE_NAME="${1:-kernel}" # defines the package name used
+KVERREL="${2:-}"
+SUBARCH="${3:-}" # defines a specific arch for use with rh-configs-arch-prep target
+SCRIPT="$(readlink -f $0)"
+OUTPUT_DIR="$PWD"
+SCRIPT_DIR="$(dirname $SCRIPT)"
+
+# to handle this script being a symlink
+cd $SCRIPT_DIR
 
 set errexit
 set nounset
 
 control_file="config_generation"
+
+cleanup()
+{
+	rm -f config-*
+}
+
+die()
+{
+	echo "$1"
+	cleanup
+	exit 1
+}
 
 function combine_config_layer()
 {
@@ -29,14 +49,27 @@ function merge_configs()
 	archvar=$1
 	arch=$(echo "$archvar" | cut -f1 -d"-")
 	configs=$2
-	name=$PACKAGE_NAME-$archvar.config
+	name=$OUTPUT_DIR/$PACKAGE_NAME-$archvar.config
 	echo -n "Building $name ... "
 	touch config-merging config-merged
+
+	# apply base first
 	for config in $(echo $configs | sed -e 's/:/ /g')
 	do
+		perl merge.pl config-base-$config config-merging > config-merged
+		if [ ! $? -eq 0 ]; then
+			die "Failed to merge base"
+		fi
+		mv config-merged config-merging
+	done
+	for config in $(echo $configs | sed -e 's/:/ /g')
+	do
+		# not all override files exist
+		test -e config-$config || continue
+
 		perl merge.pl config-$config config-merging > config-merged
 		if [ ! $? -eq 0 ]; then
-			exit
+			die "Failed to merge configs"
 		fi
 		mv config-merged config-merging
 	done
@@ -60,10 +93,12 @@ function merge_configs()
 	echo "done"
 }
 
-glist=$(find baseconfig -type d)
-dlist=$(find debugconfig -type d)
+glist=$(find base-generic -type d)
+dlist=$(find base-debug -type d)
+gllist=$(test -d generic && find generic -type d)
+dllist=$(test -d debug && find debug -type d)
 
-for d in $glist $dlist
+for d in $glist $dlist $gllist $dllist
 do
 	combine_config_layer $d
 done
@@ -86,4 +121,15 @@ do
 	fi
 done < $control_file
 
-rm -f config-*
+# A passed in kernel version implies copy to final location
+# otherwise defer to another script
+if test -n "$KVERREL"
+then
+	for i in kernel-*.config
+	do
+		NEW="$(echo $i | sed "s/$PACKAGE_NAME-$SUBARCH/$PACKAGE_NAME-$KVERREL-$SUBARCH/")"
+		mv $i $NEW
+	done
+fi
+
+cleanup
